minor comment
upgrade gradle

apply Xsource:3 inspection & quickfix everywhere

enable ""-P:kind-projector:underscore-placeholders" to be consistent with Scala 3 syntax

+++ module/core/src/test/scala/ai/acyclic/prover/commons/function/fixture/Circuits.scala
@@ -13,7 +13,7 @@ object Circuits {
   }
 
   final lazy val fn2: Long :=> Seq[Double] = :=> { v =>
-    Seq(v, v + 0.1, v + 0.2)
+    Seq(v.toDouble, v + 0.1, v + 0.2)
   }
 
   lazy val fn0Text: String = {
